fix: Shut down runtimes
